[#26] Rename: fix TabSection directiory name typo
